Adapted soldering profile based on first run
Based on the first PCB that I actually soldered
with this oven on Oct-30, 2024, I adapted
the soldering profile by increasing the
preheat time and allowing for a longer
peak time (in order to reach an actual
higher peak temperature).
Door open hint adapted accordingly.
Not yet tested, but about to be tested.